Inline tool/viewer commons and update build
Move implementations from tools/tool_common.c and tools/viewer_common.c into their respective headers (tool_common.h, viewer_common.h) and inline several common C modules into tools/fdftool.c. Update Makefile to discover tool and test sources with find, use TOOL_DEPS for tool build deps, and stop separately linking tools/tool_common.c into test binaries. Adjust tool build rule to remove the old -DTOOL_COMMON_IMPLEMENTATION/-DVIEWER_COMMON_IMPLEMENTATION flags. Other changes: tweak sprite origin in renderer/mdx/r_mdx_render.c, scale lowres_present output in tools/toolbox.c, and update VSCode task (label + command) to use "make && make run". These changes simplify build discovery and consolidate shared helper code into headers to avoid separate compilation units for small tool/viewer helpers.
